TriviaActivity Milestone3 all done
