+++ app/src/main/AndroidManifest.xml
@@ -35,7 +35,7 @@
             android:windowSoftInputMode="stateAlwaysHidden" />
 
         <activity
-            android:name=".HighScoresActivity"
+            android:name=".TriviaHighScoresActivity"
             android:exported="false"
             android:windowSoftInputMode="stateAlwaysHidden" />
