--- app/src/main/java/algonquin/cst2335/finalproject/Score.java
@@ -1,50 +1,0 @@
-package algonquin.cst2335.finalproject;
-
-import androidx.room.ColumnInfo;
-import androidx.room.Entity;
-import androidx.room.PrimaryKey;
-
-@Entity
-public class Score {
-
-    @ColumnInfo(name = "name")
-    private String name;
-
-
-
-    @ColumnInfo(name="score")
-    private int score;
-
-    @PrimaryKey(autoGenerate = true)
-    @ColumnInfo(name = "id")
-    public int id;
-
-    public Score(String name, int score) {
-
-        this.name = name;
-        this.score = score;
-    }
-
-    public int getScore() {
-        return score;
-    }
-
-    public void setScore(int score) {
-        this.score = score;
-    }
-
-    public String getName() {
-        return name;
-    }
-
-    public void setName(String name) {
-        this.name = name;
-    }
-    public int getId() {
-        return id;
-    }
-
-    public void setId(int id) {
-        this.id = id;
-    }
-}
