+++ app/src/main/java/algonquin/cst2335/finalproject/TriviaActivity.java
@@ -3,15 +3,12 @@ package algonquin.cst2335.finalproject;
 import androidx.annotation.NonNull;
 import androidx.appcompat.app.AlertDialog;
 import androidx.appcompat.app.AppCompatActivity;
-import androidx.recyclerview.widget.LinearLayoutManager;
-import androidx.recyclerview.widget.RecyclerView;
 
-import android.content.Context;
+
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.graphics.Color;
 import android.os.Bundle;
-import android.preference.PreferenceManager;
 import android.text.TextUtils;
 import android.util.Log;
 import android.view.Menu;
@@ -33,32 +30,58 @@ import org.json.JSONException;
 import org.json.JSONObject;
 
 import java.util.ArrayList;
-import java.util.List;
 
-import algonquin.cst2335.finalproject.databinding.ActivityBearBinding;
-import algonquin.cst2335.finalproject.databinding.ActivityCurrencyBinding;
 import algonquin.cst2335.finalproject.databinding.ActivityTriviaBinding;
-
+/**
+ * The TriviaActivity class represents the main activity for the Trivia Quiz application.
+ * This activity allows users to play a trivia quiz game with questions fetched from an external API.
+ * Users can choose the quiz category, specify the number of questions, answer the questions, and view their score at the end.
+ * They can also navigate to other activities like AviationActivity, CurrencyActivity, and BearActivity.
+ * The TriviaActivity class extends the AppCompatActivity class and implements various methods for UI interaction and handling quiz logic.
+ *
+ * The activity layout is defined in the activity_trivia.xml file using the ActivityTriviaBinding class.
+ * The trivia quiz questions are fetched from the Open Trivia Database API using Volley library for network requests.
+ * The user's score is saved to a local SQLite database using the TriviaScoreDatabaseHelper class.
+ * The HighScoresActivity displays the top scores from the database.
+ *
+ * Note: For the application to work correctly, it requires an internet connection to fetch questions from the API.
+ * Also, the TriviaScoreDatabaseHelper class handles the SQLite database operations for saving and querying scores.
+ *
+ * @see AppCompatActivity
+ * @see JsonObjectRequest
+ * @see Volley
+ * @see TriviaScore
+ * @see TriviaScoreDatabaseHelper
+ */
 public class TriviaActivity extends AppCompatActivity {
 
+    /**
+     * View binding variables
+     */
     ActivityTriviaBinding triviaBinding;
     TextView totalQuestionsTextView;
     TextView questionTextView;
     Button ansA, ansB, ansC, ansD;
     Button submitBtn;
-    private RecyclerView recyclerView;
-    private TriviaList adapter;
+    /**
+     * Request queue for making API calls using Volley library
+     */
     protected RequestQueue queue = null;
-
+    /**
+     *  Variables to store information about the current question and game state
+     */
     private String correctAnswer;
     private JSONArray results;
-
-
     int score = 0;
     /*int totalQuestion = TriviaActivityQA.question.length;*/
     int currentQuestionIndex = 0;
     String selectedAnswer = "";
 
+    /**
+     * Called when the activity is created. Initializes the views, sets up click listeners,
+     * and handles user interactions with the quiz.
+     * @param savedInstanceState The saved instance state of the activity (not used in this case).
+     */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
 
@@ -99,9 +122,6 @@ public class TriviaActivity extends AppCompatActivity {
             builder.setMessage("Are you sure you want to choose choose this number of questions?");
             builder.setPositiveButton("Yes", (dialog, click) -> {
                 String userInput = triviaBinding.inputEditText.getText().toString();
-
-                Toast.makeText(this, "Number of questions saved!", Toast.LENGTH_SHORT).show();
-
                 Snackbar.make(triviaBinding.getRoot(), "Your question number is: " + userInput, Snackbar.LENGTH_LONG)
                         .show();
 
@@ -112,7 +132,6 @@ public class TriviaActivity extends AppCompatActivity {
             builder.setNegativeButton("No", (dialog2, click2) -> {});
             builder.create().show();
         });
-
 
         triviaBinding.buttonSports.setText(getString(R.string.sports));
         triviaBinding.buttonSports.setOnClickListener( click -> {
@@ -129,12 +148,10 @@ public class TriviaActivity extends AppCompatActivity {
                     },
                     (error) -> {
                         int i = 0;
-                        // Handle error if necessary
-                    });            // call this for error
-            queue.add(request); //send request to server
 
+                    });
+            queue.add(request);
         });
-
 
 
         triviaBinding.buttonGeography.setText(getString(R.string.geography));
@@ -153,11 +170,8 @@ public class TriviaActivity extends AppCompatActivity {
                     },
                     (error) -> {
                         int i = 0;
-                        // Handle error if necessary
-                    });            // call this for error
-            queue.add(request); //send request to server
-
-
+                    });
+            queue.add(request);
         });
         triviaBinding.ansA.setOnClickListener(new View.OnClickListener() {
             @Override
@@ -203,10 +217,13 @@ public class TriviaActivity extends AppCompatActivity {
                 }
             }
         });
-
-
     }
 
+    /**
+     * Fetches the next trivia question and updates the UI to display the question and answer choices.
+     * If there are no more questions, it displays the final score and prompts the user to enter their name.
+     * @param currentQuestionIndex The index of the current question in the results array.
+     */
     private void fetchNextQuestion(int currentQuestionIndex) {
         try {
             if (currentQuestionIndex < results.length()) {
@@ -226,12 +243,8 @@ public class TriviaActivity extends AppCompatActivity {
                 for (int k = 1; k < 4; k++) {
                     choices[k] = incorrectTexts.get(k - 1);
                 }
-
                 runOnUiThread(() -> {
-
-//                    totalQuestionsTextView.setText("Total Question: " + results.length());
                     totalQuestionsTextView.setText(getString(R.string.total_question) + ": " + results.length());
-
                     questionTextView.setText(questionString);
                     ansA.setText(choices[0]);
                     ansB.setText(choices[1]);
@@ -242,15 +255,12 @@ public class TriviaActivity extends AppCompatActivity {
                     ansB.setBackgroundColor(Color.WHITE);
                     ansC.setBackgroundColor(Color.WHITE);
                     ansD.setBackgroundColor(Color.WHITE);
-
                     Log.d("TriviaActivity", "Next question fetched. Index: " + currentQuestionIndex);
                 });
             } else {
-                // Handle the end of trivia (no more questions)
                 runOnUiThread(() -> {
                     Toast.makeText(this, "End of trivia. Your score: " + score, Toast.LENGTH_LONG).show();
                     showEndOfTriviaDialog();
-
                     Log.d("TriviaActivity", "End of trivia. Score: " + score);
                 });
             }
@@ -259,18 +269,25 @@ public class TriviaActivity extends AppCompatActivity {
         }
     }
 
+    /**
+     * Handles the user's choice button click by highlighting the selected answer and storing it.
+     * @param clickedButton The button representing the user's selected answer.
+     */
     private void handleChoiceButtonClick(Button clickedButton) {
         triviaBinding.ansA.setBackgroundColor(Color.WHITE);
         triviaBinding.ansB.setBackgroundColor(Color.WHITE);
         triviaBinding.ansC.setBackgroundColor(Color.WHITE);
         triviaBinding.ansD.setBackgroundColor(Color.WHITE);
-
         selectedAnswer = clickedButton.getText().toString();
         clickedButton.setBackgroundColor(Color.MAGENTA);
     }
 
-
-
+    /**
+     * Called to create the options menu of the activity.
+     * Inflates the app_menu.xml layout as the options menu.
+     * @param menu The menu to be created.
+     * @return True if the menu is created successfully, false otherwise.
+     */
     @Override
     public boolean onCreateOptionsMenu(Menu menu) {
         super.onCreateOptionsMenu(menu);
@@ -278,6 +295,12 @@ public class TriviaActivity extends AppCompatActivity {
         return true;
     }
 
+    /**
+     * Handles the selection of items in the options menu.
+     * Displays appropriate dialogs or navigates to other activities based on the user's selection.
+     * @param item The selected menu item.
+     * @return True if the menu item is handled successfully, false otherwise.
+     */
     public boolean onOptionsItemSelected(@NonNull MenuItem item) {
         super.onOptionsItemSelected(item);
         AlertDialog.Builder builder = new AlertDialog.Builder(TriviaActivity.this);
@@ -320,7 +343,8 @@ public class TriviaActivity extends AppCompatActivity {
             builder.setNegativeButton(getString(R.string.no),((dialog, click) -> {}));
             builder.create().show();
         } else if (item.getItemId() == R.id.contact) {
-            //    Snackbar.make(showAmount, getString(R.string.contact_msg), Snackbar.LENGTH_LONG).show();
+            Snackbar.make(triviaBinding.getRoot(), getString(R.string.contact_msg), Snackbar.LENGTH_LONG)
+                    .show();
         } else if (item.getItemId() == R.id.about) {
             Toast.makeText(this, getString(R.string.about_msg), Toast.LENGTH_LONG).show();
         }
@@ -364,32 +388,35 @@ public class TriviaActivity extends AppCompatActivity {
         builder.show();
     }
 
+    /**
+     * Saves the user's name and score to the database.
+     * @param name The user's name.
+     * @param score The user's score in the trivia game.
+     */
     private void saveScoreToDatabase(String name, int score) {
-        // Save the user's name and score to the database (use your database implementation)
-        // For example, you can use SQLiteOpenHelper or Room database
-        // Here, we assume you have a ScoreDatabaseHelper class to handle database operations
-        ScoreDatabaseHelper databaseHelper = new ScoreDatabaseHelper(this);
-        databaseHelper.addScore(new Score(name, score));
+        TriviaScoreDatabaseHelper databaseHelper = new TriviaScoreDatabaseHelper(this);
+        databaseHelper.addScore(new TriviaScore(name, score));
     }
 
+    /**
+     * Shows the HighScoresActivity to display the top scores of previous trivia games.
+     */
     private void showHighScoresActivity() {
         // Launch the HighScoresActivity
         Log.d("TriviaActivity", "Launching HighScoresActivity");
-        Intent highScoresIntent = new Intent(TriviaActivity.this, HighScoresActivity.class);
+        Intent highScoresIntent = new Intent(TriviaActivity.this, TriviaHighScoresActivity.class);
         startActivity(highScoresIntent);
     }
 
-
+    /**
+     * Checks if a name already exists in the database, indicating a duplicate entry.
+     * @param name The name to check for duplicates.
+     * @return True if the name already exists in the database, false otherwise.
+     */
     private boolean isNameAlreadyExist(String name) {
-        // Implement a function to check if the name already exists in the database
-        // You should query your database here to check if the name already exists
-        // Return true if the name already exists, false otherwise
-        // For example, you can use SQLiteOpenHelper or Room database to check for duplicates
-        // Here, we assume you have a ScoreDatabaseHelper class to handle database operations
-        ScoreDatabaseHelper databaseHelper = new ScoreDatabaseHelper(this);
+
+        TriviaScoreDatabaseHelper databaseHelper = new TriviaScoreDatabaseHelper(this);
         return databaseHelper.isNameAlreadyExist(name);
     }
-
-
 
 }
