+++ app/src/main/java/algonquin/cst2335/finalproject/TriviaHighScoresActivity.java
@@ -12,20 +12,39 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.List;
 import java.util.concurrent.Executor;
 import java.util.concurrent.Executors;
-
-public class HighScoresActivity extends AppCompatActivity {
-
+/**
+ * An activity that displays the top 10 high scores in a RecyclerView.
+ * It allows users to see the high scores, delete scores, and view the details of individual scores.
+ */
+public class TriviaHighScoresActivity extends AppCompatActivity {
+    /**
+     * The RecyclerView used to display the list of high scores.
+     */
     private RecyclerView recyclerView;
-    private HighScoresAdapter adapter;
-    private List<Score> highscoreList;
+    /**
+     * The adapter for the RecyclerView that handles the high score list.
+     */
+    private TriviaHighScoresAdapter adapter;
+    /**
+     * The list of TriviaScore objects representing the top 10 high scores.
+     */
+    private List<TriviaScore> highscoreList;
+    /**
+     * The database helper instance for handling high scores database operations.
+     */
+    TriviaScoreDatabaseHelper databaseHelper = new TriviaScoreDatabaseHelper(this);
 
-    ScoreDatabaseHelper databaseHelper = new ScoreDatabaseHelper(this);
-
-    private HighScoresAdapter.OnItemClickListener itemClickListener = new HighScoresAdapter.OnItemClickListener() {
-
-        // Update the existing onItemClick method to show a dialog before deleting
-        public void onItemClick(Score score) {
-            AlertDialog.Builder builder = new AlertDialog.Builder(HighScoresActivity.this);
+    /**
+     * The item click listener for handling item click events in the RecyclerView.
+     * It shows a dialog to confirm score deletion or displays the details of the selected score.
+     */
+    private TriviaHighScoresAdapter.OnItemClickListener itemClickListener = new TriviaHighScoresAdapter.OnItemClickListener() {
+        /**
+         *  Update the existing onItemClick method to show a dialog before deleting
+         * @param score The database table of user and score
+         */
+        public void onItemClick(TriviaScore score) {
+            AlertDialog.Builder builder = new AlertDialog.Builder(TriviaHighScoresActivity.this);
             builder.setTitle("Confirm Deletion");
             builder.setMessage("Are you sure you want to delete this score or see the detail of the item?");
             builder.setPositiveButton("Yes", new DialogInterface.OnClickListener() {
@@ -42,26 +61,13 @@ public class HighScoresActivity extends AppCompatActivity {
             });
             builder.show();
         }
-
-
-
-
-//        @Override
-//        public void onItemClick(Score score) {
-//            // Handle the item click here, similar to the original onItemClick method
-//            Log.d("HighScoresActivity", "Delete score with id: " + score.getName());
-//            Executor thread = Executors.newSingleThreadExecutor();
-//            thread.execute(() -> {
-//                databaseHelper.deleteScore(score.getName());
-//
-//                runOnUiThread(() -> {
-//                    highscoreList.remove(score);
-//                    adapter.notifyDataSetChanged();
-//                });
-//            });
-//        }
     };
 
+    /**
+     * Called when the activity is created.
+     * Sets up the RecyclerView to display the high scores and sets the adapter.
+     * @param savedInstanceState The saved instance state of the activity (not used in this case).
+     */
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -73,6 +79,9 @@ public class HighScoresActivity extends AppCompatActivity {
 
     }
 
+    /**
+     * Sets up the RecyclerView to display the high scores and sets the adapter.
+     */
     private void setupRecyclerView() {
         recyclerView = findViewById(R.id.recyclerViewHighScores);
         recyclerView.setLayoutManager(new LinearLayoutManager(this));
@@ -82,7 +91,7 @@ public class HighScoresActivity extends AppCompatActivity {
 
         // Create the adapter with the high score list and the item click listener if it doesn't exist
         if (adapter == null) {
-            adapter = new HighScoresAdapter(highscoreList, itemClickListener);
+            adapter = new TriviaHighScoresAdapter(highscoreList, itemClickListener);
             recyclerView.setAdapter(adapter);
         } else {
             // Update the adapter's dataset with the new highscoreList
@@ -91,8 +100,11 @@ public class HighScoresActivity extends AppCompatActivity {
         }
     }
 
-
-    private List<Score> fetchTop10Highscore() {
+    /**
+     * Fetches the top 10 high scores from the database.
+     * @return The list of TriviaScore objects representing the top 10 high scores.
+     */
+    private List<TriviaScore> fetchTop10Highscore() {
         // Fetch the top 10 high score from the database (use your database implementation)
         // For example, you can use SQLiteOpenHelper or Room database
         // Here, we assume you have a ScoreDatabaseHelper class to handle database operations
@@ -100,7 +112,11 @@ public class HighScoresActivity extends AppCompatActivity {
         return databaseHelper.getTop10score();
     }
 
-    private void deleteScore(Score score) {
+    /**
+     * Deletes the specified score from the database and updates the RecyclerView.
+     * @param score The TriviaScore object representing the score to be deleted.
+     */
+    private void deleteScore(TriviaScore score) {
         Log.d("HighScoresActivity", "Delete score with id: " + score.getName());
         Executor thread = Executors.newSingleThreadExecutor();
         thread.execute(() -> {
@@ -113,9 +129,11 @@ public class HighScoresActivity extends AppCompatActivity {
         });
     }
 
-
-
-    private void showItemDetailFragment(Score score) {
+    /**
+     * Displays the details of the selected score in a fragment.
+     * @param score The TriviaScore object representing the selected score.
+     */
+    private void showItemDetailFragment(TriviaScore score) {
         // Create the fragment instance and pass the selected Score as an argument
         TriviaDetailsFragment fragment = new TriviaDetailsFragment(score);
         getSupportFragmentManager().beginTransaction()
@@ -123,5 +141,4 @@ public class HighScoresActivity extends AppCompatActivity {
                 .addToBackStack(null)
                 .commit();
     }
-
 }
