--- app/src/main/java/algonquin/cst2335/finalproject/TriviaList.java
@@ -1,50 +1,0 @@
-package algonquin.cst2335.finalproject;
-
-
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import android.widget.TextView;
-
-import androidx.annotation.NonNull;
-import androidx.recyclerview.widget.RecyclerView;
-
-import java.util.List;
-
-public class TriviaList extends RecyclerView.Adapter<TriviaList.ViewHolder>{
-    private List<String> items;
-
-    public TriviaList(List<String> items) {
-        this.items = items;
-    }
-    public static class ViewHolder extends RecyclerView.ViewHolder {
-        public TextView textView;
-
-        public ViewHolder(View itemView) {
-            super(itemView);
-            textView = itemView.findViewById(android.R.id.text1);
-        }
-    }
-    @NonNull
-    @Override
-    public TriviaList.ViewHolder onCreateViewHolder(@NonNull ViewGroup parent, int viewType) {
-        LayoutInflater inflater = LayoutInflater.from(parent.getContext());
-        View itemView = inflater.inflate(android.R.layout.simple_list_item_1, parent, false);
-        return new ViewHolder(itemView);
-    }
-
-    @Override
-    public void onBindViewHolder(@NonNull TriviaList.ViewHolder holder, int position) {
-        String item = items.get(position);
-        holder.textView.setText(item);
-    }
-
-    public void setItems(List<String> items) {
-        this.items = items;
-        notifyDataSetChanged();
-    }
-    @Override
-    public int getItemCount() {
-        return items.size();
-    }
-}
