+++ app/src/main/res/layout/trivia_fragment.xml
@@ -2,7 +2,7 @@
 <androidx.constraintlayout.widget.ConstraintLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    android:background="@color/white"
+    android:background="@color/yellow"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
 
@@ -11,9 +11,9 @@
         android:id="@+id/textItemID"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="84dp"
-        android:layout_marginTop="84dp"
+        android:layout_marginTop="80dp"
         android:text="Item ID Goes Here"
+        android:textSize="32dp"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent" />
 
@@ -21,8 +21,9 @@
         android:id="@+id/textItemName"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginTop="152dp"
+        android:layout_marginTop="104dp"
         android:text="Item Name Goes here"
+        android:textSize="32dp"
         app:layout_constraintStart_toStartOf="@+id/textItemID"
         app:layout_constraintTop_toBottomOf="@+id/textItemID" />
 
@@ -32,6 +33,7 @@
         android:layout_height="wrap_content"
         android:layout_marginTop="112dp"
         android:text="Item Score goes here"
+        android:textSize="32dp"
         app:layout_constraintStart_toStartOf="@+id/textItemName"
         app:layout_constraintTop_toBottomOf="@+id/textItemName" />
